chip/stm32/usb: Fix resume handler in usb_wake
In the previous version of the code, we were counting ESOF (every
ms) to figure out when to clear the resume bit, and then using ESOF
counts again to wait for the USB interface to resume.

This missed the fact that ESOF (expected SOF) interrupts are not
triggered when SOF are actually received.

Somehow, this worked fine in most cases, except that sometimes the
last ESOF would race with the RXDP/RXDM going to state 2, and we
would not know that the resume completed successfully.

Let's also count SOF interrupts, and also take a received SOF as an
indicator of a successful resume.

While we're at it, trim down USB debugging messages, and add a line
when resume is successful.

BRANCH=none
BUG=b:35587173
TEST=Connect hammer, force autosuspend using:
     DEVICE=$(dirname $(grep 5022 /sys/bus/usb/devices/*/idProduct))
     echo 500 > $DEVICE/power/autosuspend_delay_ms
     echo auto > $DEVICE/power/control
     Add debugging in hammer code, make sure that usb_wake_done goes
     back to 1 after resume succeeds.

Change-Id: I206c9c6b3066a3a337b6bd2370c9d0c6a9e2396c