ec: Execute EC code on QEMU
Using QEMU to run Embedded Controller (EC) code greatly simplifies
and lowers the learning curve of noobies because no hardware and
device programmers are needed. This method of running EC code is
also useful to the experienced developer by enabling them to write
code for virtual hardware and to test the code under different
memory configurations. For example, code could be developed for
a new EC before the EC was designed into a device.

The LM4 chip was ported to the LM3.

BUG=None
BRANCH=None
TEST=`make -j buildall`

Signed-off-by: Sam Hurst <shurst@chromium.org>
Change-Id: I92b1eb35966e76041251f76302e908aed5c82ab3