Update patch set 6
Patch Set 6:

> Patch Set 6: Code-Review-1
> 
> This seems like the wrong way to solve this, and will add overhead in both code size and runtime for every platform to solve a problem on only one platform.

Yeah, I thought that isn't a good way to do it as well. But I couldn't think of a better way.
> 
> How do you re-enable the touch sensor controller interrupt when the flash operation is done?

I'm using a deferred task to reenable the interrupt after some timeout.

> 
> Could the referenced bug interfere with other things that use SPI (e.g., some accelerometers)?
That is unlikely, firmware update failed with -110 (ETIMEOUT) because the i2c bus is occupied by touch events. We need some kind of way to notify the firmware that it is being updated.

I've also considered adding a new host command to notify the firmware, but that means we need to add the call in flashrom as well...

Patch-set: 6
Reviewer: Gerrit User 1116939 <1116939@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>