Update patch set 4
Patch Set 4: Code-Review+2

(2 comments)

sorry it took me so long to get to it.

The comments are more for discussion, and could be addressed later if decided to do so.

It would be great to have a better patch description too.

Patch-set: 4
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2