util/openocd: change the _CHIPNAME of npcx ec to a generalized name
The original _CHIPNAME for npcx ec in the openocd configure file is
npcx5m5g. As we introduced more npcx ec SKUs, it is not appropriate to
keep using this name. This CL modify the _CHIPNAME to a generalized
name(npcx_ec).

BRANCH=none
BUG=none
TEST=./util/flash_ec --board=npcx7_evb; check the openocd log and make
sure the _CHIPNAME is npcx_ec now.

Signed-off-by: CHLin <CHLIN56@nuvoton.com>
Change-Id: I49d298de1e43ac29f3e5535702595ee27225ac23