CNL: Don't build CNL in tests w/o chipset tasks.
When compiling target tests for a board that defines
CONFIG_CHIPSET_CANNONLAKE, the build would fail due to the chipset code
being built without a chipset task.  This commit removes the Cannonlake
chipset code when building a test that does not have a chipset task (and
use Cannonlake).

BUG=None
BRANCH=None
TEST=make -j BOARD=zoombini tests; verify that build succeeds.

Change-Id: I15a5729db64314a8d6566be879f7c356ffa95725
Signed-off-by: Aseda Aboagye <aaboagye@google.com>