Update patch set 21
Patch Set 21:

> Patch Set 21:
> 
> > Patch Set 21:
> > 
> > > Patch Set 21: Code-Review-1
> > > 
> > > I'm seeing a case where kevin is not waking up on a lid open event, so holding off any merge until I determine cause of the lid open event not waking up kevin on occasion and make sure it's not related to this CL.
> > 
> > Hi Nick,
> > there's a knowing issue for lid open not waking on kevin:
> > https://partnerissuetracker.corp.google.com/u/1/issues/36042967
> > 
> > if that is what you saw, then it would still be ok to land this CL :)
> 
> and we can confirm it by checking the ec log, or try my patch in the issue.

I believe what I saw was a case where Lid Open was still pending in events (hadn't been cleared yet), so the subsequent event was ignored because it was deemed that nothing changed.  Thoughts were to make host_set_events() always wake AP if event is in wakeup mask regardless of whether event is already pending, but haven't been able to repro the case to confirm this is what occurred.

But I've banged on this enough at this point to feel comfortable enough to feel the issue is not specific to this new change.

Patch-set: 21
Reviewer: Gerrit User 1186109 <1186109@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0