Update patch set 3
Patch Set 3:

My understanding was that this defaults to internal pulldown and should not be pulled out to a probeable spot on production pcbs. I guess not tho. May be safer to put an internal pulldown too so it's independent of pcb design.

Patch-set: 3
Reviewer: Gerrit User 1000104 <1000104@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>