eve: Don't disable keyboard scanning in S0.
The EC shouldn't unconditionally disable keyboard scanning in S0; the
decision to ignore keyboard presses should be left up to the AP.  This
commit changes the behaviour to only disable keyboard scanning if the
chipset is not on.

BUG=b:63904951, b:63905669
BRANCH=eve
TEST=Flash eve; extend display to external monitor and make it the only
display, place DUT into tablet mode, verify that keyboard presses still
work.

Change-Id: I2bb37d5e5b628254e65076c8305ecb2987d4c1de
Signed-off-by: Aseda Aboagye <aaboagye@google.com>