eve: PROCHOT behavior fixes
1) Don't assert PROCHOT by default. This signal has an inverter
and driving it high is asserting prochot until it gets deasserted
during S3->S0 transition in power/intel_x86.c.

2) Clear PMIC power source interrupts when entering S0 to ensure
that the PMIC is not asserting PROCHOT.  This register is not cleared
by an LDO reset so it was never able to get cleared before.  I want
to ensure it is cleared on the way into S0 so I didn't put it into
pmic_init as that would only happen on an EC reset.

BUG=b:63913242
BRANCH=eve
TEST=manual testing on Eve unit

Change-Id: I5c7f7ca528160972f65bd714deea98357fa43a53
Signed-off-by: Duncan Laurie <dlaurie@google.com>