Update patch set 2
Patch Set 2:

CCD now is the only way to control write protect.

It must be possible to do this without using the CCD cable.

So, we need to be able to ccdopen and toggle WP via cr50tool.  And that all needs to keep working across deep sleep.

CCD is more than just the USB endpoints (which yes, go away when the cable is unplugged).  It's all of the case-closed debugging configuration.

See "Enabling CCD Must Not Require Opening the Case Or Specialized Tools" in go/cr50-ccd-wp

Patch-set: 2
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>