Update patch set 7
Patch Set 7:

> Uploaded patch set 7.

This version is more or less directly what GCC is doing, only that it pushes r3 instead of r4 to keep the stack 64-bit aligned.

The spirit here is to bring a minimal diff compared to the current situation, so the ASM code is not optimized. Optimizations are out of the scope of this patch and should be made in follow-up patches if desired.

The goal is to be able to merge this without a very long discussion, as being able to build cros-ec with newer GCC versions became more of a priority.

This was additionally tested on minnie and my broken elm unit.

Patch-set: 7
Reviewer: Gerrit User 1133285 <1133285@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>