Update patch set 6
Abandoned

Abandoning this, as the problem was never witnessed on cortex-m. This was done as a cautionary measure, but I don't think it's worth the trouble anymore. GCC doesn't add any tailing pop after __switchto in any version I could test, so we are good.

Patch-set: 6
Status: abandoned
Tag: autogenerated:gerrit:abandon