eve: Remove the LIMIT_POWER config options
These config options change the behavior of charge_prevent_power_on
and ignore the minimum battery percentage for booting.  Since we
do not have any AP code to actually handle this state we don't
want it to always boot the AP or it might brown out with a battery
that is critically low.

BUG=b:63957122
BRANCH=eve
TEST=manual testing with low battery to ensure it does not always
attempt to boot the AP.

Change-Id: Ic7be3f4d07570e00ea3afa1124af083da0e789c3
Signed-off-by: Duncan Laurie <dlaurie@google.com>