eve: Cut off critically low battery
Instead of going to hibernate when the battery is critically low
we should cut off power entirely.

Even with the PMIC shut down the H1 chip consumes more power than
is healthy when the battery is already critically low, depleting it
to dangerously low voltage levels faster than it should.

BUG=b:63957122
BRANCH=eve
TEST=manual testing on Eve with critically low battery to ensure that
it is cut off instead of going into hibernate.

Change-Id: I2eaba3623385a4eb3daa39a2fa8aa08d3ec6366e
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/582543
Reviewed-by: Furquan Shaikh <furquan@chromium.org>