prevent_power_on: Wait for charge current to be initialized
When allowing an unlocked system to automatically power up we
may attempt to boot before the charger is ready to supply enough
power and get stuck in a reset loop.

By adding a final check to the charge_prevent_power_on() function
to ensure that charge_manager_get_charger_current() is returning
a valid value instead of CHARGE_CURRENT_UNINITIALIZED then the
board is able to reliably boot.

BUG=b:63957122
BRANCH=eve
TEST=manual testing on Eve to reliably boot without a battery

Change-Id: Ie7b5db49eb3a36dc3197731a383fba26db839610
Signed-off-by: Duncan Laurie <dlaurie@google.com>