g: provide a function to check image and board ID match
The use of this function will allow limiting certain features to
certain H1 chips based on the Board ID contents programmed in Info1.

The code checking if a certain feature is allowed to run would provide
the header_board_id structure with the contents matching the board
where the feature is allowed. The new function would compare the
passed in structure against the current Info1 contents and return True
or False depending if the passed in structure is matching the chip the
code is running on.

BRANCH=cr50
BUG=b:62294740
TEST=with the rest of the patches applied verified the ability to
     limit certain code features to certain chips

Change-Id: Ibe57e0045256b99e8580f2e08e17ae6eef8dd5a9
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>