Update patch set 2
Patch Set 2: Code-Review-1

This makes me nervous.

The point of board_id locking images is to make the decision as to what features are allowed at build/signing time, and then have a strong check at boot time (ideally in the RO image).

With this proposal, someone signing a prod image needs to check all the places where a board_id is dynamically checked at runtime.

Unless the check is repeated several times, it's susceptible to clock glitching every time it's made.  That may be exploitable - for example, if I can request some operation repeatedly, and a failure doesn't trigger reboot, I'll eventually be able to enable it on a disallowed board.

Maybe not horrible for U2F, but we wouldn't want to gate CCD V1 this way.

It would be good to add this to the board ID locking design doc and discuss this change with a few more folks from security.

Patch-set: 2
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1