cr50: protect EC/AP reset commands using CCD V1
The sysrst and ecrst commands are now protected by the RebootECAP
capability.  They can print the state of the reset lines when the
capability is not allowed, but can only change the state or pulse the
reset lines when the capability is allowed.

See go/cr50-ccd-wp for more information.

BUG=b:62537474
BRANCH=none
TEST=manual with CR50_DEV=1 build:
     ccdoops --> reset CCD config and go to Opened state
     ecrst pulse --> works
     sysrst pulse --> works

     ccdunlock
     ecrst pulse --> access denied
     sysrst pulse --> access denied
     ecrst --> prints state
     sysrst --> prints state

     ccdoops
     ccdset rebootecap unlesslocked

     ccdunlock
     ecrst pulse --> works
     sysrst pulse --> works

Change-Id: Ia9ebe67bdc1e85129051caf94f20fb2fb84b76da
Signed-off-by: Randall Spangler <rspangler@chromium.org>