eve: Check for XCHG|XDSG bits for determining battery presence
When the battery is cutoff by the EC via I2C sb message, when the
system is is booted again (after applying external power) the EC does
not detect the battery and therefore batt.is_present is set to BP_NO
and this kicks in the 1 second delay to prevent the AP from turning on
right away. However, if the battery is cutoff via H1, when the
EC is powered back up, it would detect the battery present and the 1
second delay is not activated, even though the battery isn't ready to
supply power to the system.

This CL adds a check to the eve specific battery_is_present()
function. The bits XCHG and XDSG are checked for the condition of both
being set to also indicate batt.is_present = BP_NO which triggers the
1 second delay to the AP power up and makes recovering from both
methods of battery cutoff have equivalent paths.

BUG=b:63360549
BRANCH=eve
TEST=manual Iniated battery cutoff in the following 3 methods:
1) issue 'cutoff' command on EC console
2) discharge the battery so that the EC initateds the battery shutdown
sequence when the battery voltage drops too low.
3) H1 generated battery cutoff

For methods 1 and 2, the following console output is found:
[0.006188 battery not found]
For all 3 methods the next two console outputs are found.
[1.486592 battery woke up]
[2.487140 battery will now allow booting]

In addition to these tests, also put the EC into hibernate, then using
an external discharge circuit, discharged the battery below v_min
(~5.0 V), then tested that the system powered back up and the battery
resumed charging.

Change-Id: I64747cab406ef194ee546c7520ae6479b3a8301d
Signed-off-by: Scott Collyer <scollyer@google.com>
Reviewed-on: https://chromium-review.googlesource.com/585837
Tested-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@google.com>
Commit-Queue: Duncan Laurie <dlaurie@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2320247
Commit-Queue: Patryk Duda <pdk@semihalf.com>
Tested-by: Patryk Duda <pdk@semihalf.com>
Reviewed-by: Jett Rink <jettrink@chromium.org>