Update patch set 1
Patch Set 1:

So I verified by disabling below config the issue is not reproducible, which makes sense because it will keep the pd_comm_locked.
+++ b/board/poppy/board.h
@@ -20,7 +20,6 @@
  * Allow dangerous commands.
  * TODO: Remove this config before production.
  */
-#define CONFIG_SYSTEM_UNLOCKED

So since this CONFIG will be removed before production, shouldn't be an issue to the end user.
Question is do we still want to fix when in development phase?

Patch-set: 1
Reviewer: Gerrit User 1129265 <1129265@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>