Disable sysjump when booting without a battery
In order to be able to disable write protect we need to be able to
boot without a battery.  Currently this is difficult because the
Type-C port is reset and the AP has a very high chance of resetting
or losing power.

If we check for the lack of battery, as well as ensure that the system
is not locked, then we can skip the sysjump and be able to boot to
the OS so software WP can be disabled.

BUG=b:64177860
BRANCH=eve
TEST=Remove the battery and successfully boot the AP with GBB=0

Change-Id: I88772f6502cbbfd02ef1c1f55fa3040c75f8cd4b
Signed-off-by: Duncan Laurie <dlaurie@google.com>