Update patch set 6
Patch Set 6:

If this works well enough for the KX022, maybe we can land it as is on the oak branch. As far as I'm aware, no boards built from that branch use the KXCJ9. This way we can let the factory testing continue. Before cherry-picking the change elsewhere, we'll need to figure out what's going on.

Patch-set: 6
Reviewer: Gerrit User 1125984 <1125984@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>