presubmit_check: Don't include flash_ec.
One of the items that the presubmit check checks is that if any source
file has been modified, the unit tests are run again.  However, flash_ec
is included in the list of files.  This commit removes flash_ec from the
candidates because it's not run in any unit tests anyways and it's a
shell script that is not compiled.

Additionally, ignore changes to the presubmit_check.sh itself.

BUG=None
BRANCH=None
TEST=modify flash_ec, verify that presubmit check doesn't complain about
not running make buildall.

Change-Id: I473f269c7f0bba9be3216fcec39f266d896fb908
Signed-off-by: Aseda Aboagye <aaboagye@google.com>