Fizz: Reject power role swap
This patch makes Fizz refuse to swap power roles. If it boots on BJ,
the type-c port will be only work as a source. If it boots on type-c,
it'll stay as a sink.

Without this patch, PR swap (SRC->SNK) causes Fizz to brown out.

BUG=b:64277216
BRANCH=none
TEST=Boot on BJ. Verify Pixel XL is charged by type-c port. Initiate
PR swap from Pixel. Verify PR swap is rejected. Boot on type-c.

Change-Id: I8c3161ee9873e52007d4236fc762bff652f3ce9c
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>