ec: Add a task information dump tool for the stack analyzer.
Add a tool to dump task information.
Modified the stack analyzer to get task information from the tool.
Show allocated stack sizes of tasks in the stack analyzer.

To get the all task information (including the allocated stack size),
A utility tool is added and compiled with the board to export all
configurations of the tasklist.

BUG=chromium:648840
BRANCH=none
TEST=make BOARD=elm utils && build/elm/util/ec_dump_taskinfo -s RW
     extra/stack_analyzer/stack_analyzer_unittest.py
     make BOARD=elm && extra/stack_analyzer/stack_analyzer.py \
         --objdump=arm-none-eabi-objdump \
         --addr2line=arm-none-eabi-addr2line \
         ./build/elm/RW/ec.RW.elf ./build/elm/RW/ec.RW.taskinfo
     make BOARD=${BOARD} SECTION=${SECTION} analyzestack

Change-Id: Ia5bd555b99b00edc71af20f1bfd3d0da838d63bd
Signed-off-by: Che-yu Wu <cheyuw@google.com>