cr50: Refactor device state machines
WORK IN PROGRESS; DO NOT SUBMIT

The servo, AP, and EC state machines are not quite similar enough to
use a common device state handler for all of them.  That scatters the
state code for each device between multiple files.  This wasn't too
bad before, but it's about to get more complex when the CCD V1 state
machine enters the picture in the next CL.

Refactor each of the state machines into its own file, so that all of
the state transitions are handled in one place per device.

Consolidate all of the code for changing port settings into rdd.c,
where it can be triggered by a CCD config change or by
rdd_update_state() called from other device state machines.

This also fixes some problems where disabling bitbanging would leave
UART TX enabled, even if it should have been disabled because the EC
is off or we don't have CCD permission to write to it.

BUG=b:62537474
BRANCH=cr50
TEST=manual, on the test board
	Pull EC_DETECT high and low and see that the device
	connects/disconnects properly.

	Repeat for AP_DETECT and SERVO_DETECT.

	With SERVO_DETECT disconnected, enable UART bit banging and
	gpioset the TX line to 0.  Then connecting and disconnecting a
	pullup to the signal does not change the servo state.

Change-Id: I641f7ab2354570812e3fb37b470de32e5bd10db7
Signed-off-by: Randall Spangler <rspangler@chromium.org>