eve: Modify LED behavior to show double tap patterns when charging
Previous LED behavior was to have only the LED on side that was
charging show white or green. The other side LED was always off.

This CL modifies the behavior so that when double tap patterns are
allowed and a charger is connected to display the double tap pattern
on the opposite side LED which was previously off.

In addition, when the device is charging, if the power level is such
that AP can't boot, the opposite side LED will display the BLINK_RED
pattern continously as it would duirng a double tap event.

In order to support the ability to have different LED patterns
simultaneously, the LED code was refactored so that static variables
used to manage the patterns and varialbes required to manage color
gradient transitions were grouped together in to a static led
descriptor. In addition, the led_manage_pattern and change_color
functions had to be modified to deal with each LED independently.

BUG=b:35584895
BRANCH=eve
TEST=

Change-Id: I05e1671deada761388bf898096ed7d3ae6f8da0f
Signed-off-by: Scott Collyer <scollyer@google.com>