kevin / nefario: Add comment to prevent reordering PWM IDs
As can be seen in <http://crosreview.com/610449> the IDs of PWMs are
directly referenced by the kernel device tree on gru / kevin / nefario
boards (technically also in bob and scarlet-rev0, but neither of those
is supported by mainline EC code).  Because of the kernel reference
it's not a great idea to change the IDs.

Let's add a comment so people don't change the ordering / IDs
accidentally.

NOTE: we can have a big argument here about whether it would be OK to
change numbering _and_ change the numbering in the kernel at the same
time.  We can talk about whether the device tree is a binary and about
whether this constitutes an ABI since Chromebooks don't ship the
device tree binary separate from the kernel.  We can talk about all
those things.  ...but we won't.  There's no good reason to change the
ID, so just don't do it.

BRANCH=None
BUG=b:63537905
TEST=None

Change-Id: Iacdedc1c91e583183034d15e99735a470d6e0951
Signed-off-by: Douglas Anderson <dianders@chromium.org>