usb_pd_protocol: Req SNK Cap if not received yet.
In the SRC_READY state, we'll only request sink caps if we haven't
received them yet and only if its the first transition to the state.
However, we also don't send any PD traffic in that state if there's an
incoming message in order to prevent a collision.  This could create a
scenario where upon entry to the SRC_READY state, a message is incoming.
When this occurs, we never request the sink caps.

This commit simply removes the condition that we may only request sink
caps on the first transition to the SRC_READY state.

BUG=b:64037926
BRANCH=gru
TEST=Flash kevin; Connect to a DR port partner; Verify that sink caps
are requested even after the first transition to the SRC_READY state.

Change-Id: Ia1393c7a866a40a2b588b12fbb318235341daa2b
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Commit-Id: 8d6da80f34cab49cd773e462d9720a210d4ce471
Original-Change-Id: I6bc9ad01d45e6584a7a14b28806ae4872a22d98f
Original-Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/611320
Original-Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Original-Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Original-Reviewed-by: Shawn N <shawnn@chromium.org>