usb_pd_protocol: Soft Reset after VDM errors.
When there is an explicit contract in place and a VDM is sent, the
partner shall respond with at least a GoodCRC.

The USB PD 2.0 v1.3 spec says:

"A failure to see a GoodCRC Message in response to any Message within
tReceive (after nRetryCount retries), when a Port Pair is Connected, is
indicative of a communications failure resulting in a Soft Reset (see
Section 6.5.10.1)."

This commit issues a Soft Reset in the SRC_READY state when a VDM goes
unanswered.

BUG=b:64037926
BRANCH=gru
TEST=Flash kevin; Connect to a device which doesn't respond to certain
SVDMs.  Using a twinkie, verify that a soft reset is issued after
waiting for a response to the SVDM.

Change-Id: I538a2788a5e16432599e237577ae406bf3b5643f
Signed-off-by: Aseda Aboagye <aaboagye@google.com>