stm32: add nucleo-f767zi board
/!\ DO NOT SUBMIT /!\
We have enough useless Nucleo / dev boards in the tree.
Just as a proof of concept for STM32F767/769 support.

The STM32F76x is really close to the STM32F4 family, so the most concise
implementation is just using CHIP_FAMILY_STM32F4 and adding
CHIP_VARIANT_F76X.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=none
BUG=none
TEST='make BOARD=nucleo-f767 flash', the red LED is on and the green LED
turns on/off when pressing the user button, do a capture on the SPI
sensor on SPI3 interface

Change-Id: Idfcc9bb9d7c35a957e391bf8f371ad675565a9c1