WIP: Generate hashes for the touchpad FW
Based on the passed TOUCHPAD_FW parameter, the Makefile generates
hashes for the touchpad FW, in blocks of CONFIG_TOUCHPAD_VIRTUAL_SIZE.

This will allow the USB updater code to verify the integrity of
the touchpad firmware being flashed.

When no FW is provided, SHA256 full of zeros are output, which
should not match any good firmware.

BRANCH=none
BUG=b:63993173
TEST=make TOUCHPAD_FW=SA459C-1211_ForGoogleHammer_3.0.bin \
          BOARD=hammer -j

Change-Id: Ie347270aa9c00342de13489c9422e45e681b94c2
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>